data_processing/requirements.txt
pandas==2.0.0
numpy==1.24.2
requests==2.28.2
vaderSentiment==3.3.2
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
python-dotenv==1.0.0
